Update #2
Second big update.

I noticed an error when trying to equip a crossbow when a character
already had both hands full.

I then just rewrote the function for equipping weapons to be simpler and
added an option to view the weapon set and any potential conlicts before
chosing to equip it.

Armors have a similar option but the code could be simpler.

I may tackle it at a later time.